+++ CHANGELOG.md
@@ -10,6 +10,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 * Let cinit always inherit zombie processes left by its children
   ([#37](https://gitlab.com/veenj/cinit/issues/37))
 
+* Implement drop-in configuration
+  ([#36](https://gitlab.com/veenj/cinit/issues/36))
+
 ## Release [1.2.6]
 
 ### Maintenance